moving variable helpers outside of context
